server/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
transformers
torch
pandas
python-dotenv
supabase
requests